Added min improvement(per optimization round) as script parameter.
Added a optimumRange parameter to the optimizer( by setting this value to y the optimizer will stop learning until it optimum is < (y * lowestNfpValue in sample set) in case the optimum is not the optimum in the sample set) and as script parameter.
By setting it to 1 it will stay neutral.